Add ffaker dependency to template gemspec
ffaker was removed as a runtime dependency of Solidus.

We have a require for ffaker in the spec_helper. Even if the author of
the new extension does not need to use ffaker and removes that require
statement, Solidus will complain that Solidus factories need the
library.

PR removing ffaker from Solidus:
https://github.com/solidusio/solidus/pull/2163